Create homepage as well as 3 unique recipe pages
Built the structure of my recipe website. Created the index.html file as my home page which contains a heading as well as an unordered list which contains links to 3 separate recipe pages. The recipe pages each contain the recipes description, ingredients and steps on how to make the dish.